// src/gen-ts/src/models/Leave.ts
export interface Leave {
  /**
   *
   * @type {string}
   * @memberof Leave
   */
  comment?: string
  /**
   *
   * @type {Date}
   * @memberof Leave
   */
  created_at?: Date
  /**
   *
   * @type {Date}
   * @memberof Leave
   */
  updated_at?: Date
  /**
   *
   * @type {AuditUser}
   * @memberof Leave
   */
  created_by?: AuditUser
  /**
   *
   * @type {AuditUser}
   * @memberof Leave
   */
  updated_by?: AuditUser
  /**
   *
   * @type {string}
   * @memberof Leave
   */
  id?: string
  /**
   *
   * @type {CrupdateUser}
   * @memberof Leave
   */
  user?: CrupdateUser
  /**
   *
   * @type {LeaveType}
   * @memberof Leave
   */
  leave_type?: LeaveType
  /**
   *
   * @type {Date}
   * @memberof Leave
   */
  start_date?: Date
  /**
   *
   * @type {Date}
   * @memberof Leave
   */
  end_date?: Date
  /**
   *
   * @type {number}
   * @memberof Leave
   */
  duration_days?: number
  /**
   *
   * @type {LeaveStatus}
   * @memberof Leave
   */
  status?: LeaveStatus
  /**
   *
   * @type {string}
   * @memberof Leave
   */
  reason?: string
  /**
   *
   * @type {CrupdateUser}
   * @memberof Leave
   */
  approved_by?: CrupdateUser
  /**
   *
   * @type {Date}
   * @memberof Leave
   */
  approved_at?: Date
}