// src/gen-ts/src/models/CrupdateLeave.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { LeaveStatus } from './LeaveStatus'
import {
  LeaveStatusFromJSON,
  LeaveStatusFromJSONTyped,
  LeaveStatusToJSON,
  LeaveStatusToJSONTyped,
} from './LeaveStatus'

/**
 *
 * @export
 * @interface CrupdateLeave
 */
export interface CrupdateLeave {
  /**
   *
   * @type {string}
   * @memberof CrupdateLeave
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeave
   */
  user_id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeave
   */
  leave_type_id?: string
  /**
   *
   * @type {Date}
   * @memberof CrupdateLeave
   */
  start_date?: Date
  /**
   *
   * @type {Date}
   * @memberof CrupdateLeave
   */
  end_date?: Date
  /**
   *
   * @type {number}
   * @memberof CrupdateLeave
   */
  duration_days?: number
  /**
   *
   * @type {LeaveStatus}
   * @memberof CrupdateLeave
   */
  status?: LeaveStatus
  /**
   *
   * @type {string}
   * @memberof CrupdateLeave
   */
  reason?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLeave
   */
  comment?: string
}

/**
 * Check if a given object implements the CrupdateLeave interface.
 */
export function instanceOfCrupdateLeave(value: object): value is CrupdateLeave {
  return true
}

export function CrupdateLeaveFromJSON(json: any): CrupdateLeave {
  return CrupdateLeaveFromJSONTyped(json, false)
}

export function CrupdateLeaveFromJSONTyped(json: any, ignoreDiscriminator: boolean): CrupdateLeave {
  if (json == null) {
    return json
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    user_id: json['user_id'] == null ? undefined : json['user_id'],
    leave_type_id: json['leave_type_id'] == null ? undefined : json['leave_type_id'],
    start_date: json['start_date'] == null ? undefined : new Date(json['start_date']),
    end_date: json['end_date'] == null ? undefined : new Date(json['end_date']),
    duration_days: json['duration_days'] == null ? undefined : json['duration_days'],
    status: json['status'] == null ? undefined : LeaveStatusFromJSON(json['status']),
    reason: json['reason'] == null ? undefined : json['reason'],
    comment: json['comment'] == null ? undefined : json['comment'],
  }
}

export function CrupdateLeaveToJSON(json: any): CrupdateLeave {
  return CrupdateLeaveToJSONTyped(json, false)
}

export function CrupdateLeaveToJSONTyped(
  value?: CrupdateLeave | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    id: value['id'],
    user_id: value['user_id'],
    leave_type_id: value['leave_type_id'],
    start_date:
      value['start_date'] == null
        ? value['start_date']
        : value['start_date'].toISOString().substring(0, 10),
    end_date:
      value['end_date'] == null
        ? value['end_date']
        : value['end_date'].toISOString().substring(0, 10),
    duration_days: value['duration_days'],
    status: LeaveStatusToJSON(value['status']),
    reason: value['reason'],
    comment: value['comment'],
  }
}